BackEnd GymPoint

+++ src/database/migrations/20201119173405-create-student-enrollment-plans.js
@@ -16,7 +16,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       created_at: {
